push before diff

+++ Castle/SiegeEngine/Core/Rendering/Shaders/AssetShader.cs
@@ -9,7 +9,7 @@ layout(location = 2) in vec2 aTexCoord;
 layout(location = 3) in vec3 aNormal;
 layout(location = 4) in float aMaterialIndex;
 layout(location = 5) in vec3 aTangent;
-layout(location = 6) in ivec4 aBoneIDs;
+layout(location = 6) in vec4 aBoneIDs;
 layout(location = 7) in vec4 aWeights;
 out vec2 TexCoord;
 out vec3 Normal;
@@ -29,7 +29,7 @@ void main()
     float totalWeight = 0.0;
     if (uHasBones == 1) {
         for (int i = 0; i < 4; i++) {
-            int boneIndex = aBoneIDs[i];
+            int boneIndex = int(aBoneIDs[i]);
             if (boneIndex == -1) continue;
             mat4 boneTransform = uBoneTransforms[boneIndex];
             vec4 localPosition = boneTransform * vec4(aPosition, 1.0);
